Migrated to the DotSDL NuGet package.
    * Now using the NuGet package instead of git submodules (yay!).
    * Changed update frequency to hertz in order to target the latest DotSDL
      API.
    * Migrated away from the obsolete GetCanvasPointer override from before to
      the new method (changing the function pointer on the canvas object
      itself).

+++ Source/Lyon/App/IWindow.cs
@@ -3,7 +3,7 @@
     public interface IWindow
     {
         void SetSize(int width, int height);
-        void Start(int updateRate);
+        void Start(float updateRate);
         void Close();
     }
 }
